requirements.txt
openai
python-dotenv
PyPDF2
ffmpeg-python
